feat: Refactor event location form for create event
