+++ src/modules/organization/createEvent/event-location-form/EventLocationForm.vue
@@ -22,14 +22,12 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, Ref, ref, watch, reactive } from "vue";
+import { defineComponent, toRefs } from "vue";
 import FormOptionButton from "../form-option-button/FormOptionButton.vue";
 import FormSpecifyLocation from "../form-specify-location/FormSpecifyLocation.vue";
-import {
-  EventLocationFormOption,
-  UPDATE_MODEL_VALUE
-} from "@/commons/constant";
+import { UPDATE_MODEL_VALUE } from "@/commons/constant";
 import { EventLocationForm } from "@/commons/Interfaces";
+import useEventLocationForm from "./useEventLocationForm";
 
 export default defineComponent({
   name: "EventLocationForm",
@@ -45,71 +43,17 @@ export default defineComponent({
   },
   emits: [UPDATE_MODEL_VALUE],
   setup(props, context) {
-    const currentOption: Ref<EventLocationFormOption> = ref(
-      EventLocationFormOption.SPECIFY
-    );
+    const { modelValue } = toRefs(props);
 
-    const specifyEventLocation: Ref<EventLocationForm> = ref({
-      name: props.modelValue.name,
-      description: props.modelValue.description,
-      googleMapUrl: props.modelValue.googleMapUrl,
-      isOnline: props.modelValue.isOnline,
-      isValid: props.modelValue.isValid
-    });
-
-    const laterEventLocation: EventLocationForm = {
-      name: "Announce later",
-      description: "",
-      googleMapUrl: "https://www.onepass.app/",
-      isOnline: false,
-      isValid: true
-    };
-
-    const onlineEventLocation: EventLocationForm = {
-      name: "Online",
-      description: "",
-      googleMapUrl: "https://www.onepass.app/",
-      isOnline: true,
-      isValid: true
-    };
-
-    const isSpecifyOption = computed(() => {
-      return currentOption.value === EventLocationFormOption.SPECIFY;
-    });
-
-    const isLaterOption = computed(() => {
-      return currentOption.value === EventLocationFormOption.LATER;
-    });
-
-    const isOnlineOption = computed(() => {
-      return currentOption.value === EventLocationFormOption.ONLINE;
-    });
-
-    function toSpecifyOption() {
-      currentOption.value = EventLocationFormOption.SPECIFY;
-    }
-
-    function toLaterOption() {
-      currentOption.value = EventLocationFormOption.LATER;
-    }
-
-    function toOnlineOption() {
-      currentOption.value = EventLocationFormOption.ONLINE;
-    }
-
-    watch(
-      () => specifyEventLocation,
-      () => context.emit(UPDATE_MODEL_VALUE, specifyEventLocation.value),
-      { deep: true }
-    );
-
-    watch(currentOption, () => {
-      if (currentOption.value === EventLocationFormOption.SPECIFY)
-        context.emit(UPDATE_MODEL_VALUE, specifyEventLocation.value);
-      else if (currentOption.value === EventLocationFormOption.LATER)
-        context.emit(UPDATE_MODEL_VALUE, laterEventLocation);
-      else context.emit(UPDATE_MODEL_VALUE, onlineEventLocation);
-    });
+    const {
+      specifyEventLocation,
+      isSpecifyOption,
+      isLaterOption,
+      isOnlineOption,
+      toSpecifyOption,
+      toLaterOption,
+      toOnlineOption
+    } = useEventLocationForm(modelValue, context);
 
     return {
       specifyEventLocation,
